Files rename and fix for cmake
1. compile_commands.json now at top inorder for all files to link
properly.
2. added algo.cpp tp gtest so it's class can be used.
