+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "data.hpp"
-#include "header.hpp"
+#include "algo.hpp"
 #include <array>
 #include <algorithm>
